+++ src/Qiniu/CDN/FluxInfo.cs
@@ -1,5 +1,5 @@
 ﻿using System.Collections.Generic;
-
+using System;
 namespace Qiniu.CDN
 {
     /// <summary>
@@ -65,12 +65,12 @@ namespace Qiniu.CDN
             /// <summary>
             /// 国内流量数据
             /// </summary>
-            public List<int> China { get; set; }
+            public List<UInt64> China { get; set; }
 
             /// <summary>
             /// 海外流量数据
             /// </summary>
-            public List<int> Oversea { get; set; }
+            public List<UInt64> Oversea { get; set; }
         }
 
     }
